+++ tests/testthat/test-mutate.r
@@ -362,8 +362,3 @@ test_that("no utf8 invasion (#722)", {
   source("utf-8.R", local = TRUE)
 })
 
-test_that("mutate warns about unsupported attributes", {
-  d <- data.frame( x = structure( 1:10, foo = "bar" ) )
-  expect_error( d %>% group_by(x), "has unsupported attributes" )
-})
-